Support API sign in/re-send verification code
Update the `SignInsController` to support calling out to the API in order to
exchange a verification code for candidate information. The `SessionToken` is
bypassed as we only support sign in from the same device used to generate the
verification code. Updated to also issue a verification code via the API when
the feature flag is enabled.

The `GitisAuthentication` concern has been updated to support both
`Bookings::Gitis::Contact` and
`GetIntoTeachingApiClient::SchoolsExperienceSignUp`.