Update EventLogger to write via API
Instead of constructing the log line in the app this is now deferred to the
API, which accepts POST request containing a
`GetIntoTeachingApiClient::ClassroomExperienceNote`.

The class is a little misleading in that `write_later` will not queue the
request to the API, however the operation should be much faster as the request
returns immediately and is queued on the API-side now. When we pull out the
legacy code we can rename this method so its clear that the request happens
in-band.