Update ContactFetcher to work against API
The SE app has fairly complex code around getting back to the 'master' contact
record in the CRM (a contact has a 'master' when the record has been de-duped
and merged with another record).

Update to call the API instead of Dynamics directly.

Extract the `been_merged?` logic out of the `Candidate` model and into the
`ContactFetcher` - it was better placed in the `Candidate` model, but doing
this allows us to modify the logic so it also works with a
`SchoolsExperienceSignUp` model.

The `ContactFetcher` tests have been duplicated and ran against the app when
the `git_api` flag is enabled.