Test CSV exporter retrieving contacts via API
The CSV exporter uses the `ContactFetcher` to retrieve contacts associated with
placement requests prior to exporting them. Now that the `ContactFetcher` has
been updated to support retrieving contacts via the API, it makes sense to test
the exporter functionality as well when the `git_api` feature is enabled.

As the legacy code will be removed eventually I opted to duplicate the test
cases instead of extracting them to keep it DRY.