Raise a message to Sentry on unexpected direct call to CRM
If the `git_api` feature is enabled all requests should be going via the
`GetIntoTeachingApiClient`. To ensure this is the case, the existing `Dynamics`
module that handles reading/writing to the CRM currently will raise a message
to Sentry if its ever called when the `git_api` feature is on.