Update SubjectSyncJob to support git_api feature
When the `git_api` feature is enabled we want to sync via the GiT API and not
directly.

Add attribute aliases to `TeachingSubject` to achieve parity with `LookupItem`
models so that they can be used interchangeably in the sync job.

Update sync job to call out to the API when the `git_api` feature is enabled.