Update VerificationCode model to support exchange/issue
It turns out we have multiple ways a candidate can matchback; off personal
information via a registration or as part of a candidate 'sign in'.

As the candidate sign in doesn't persist the matchback data to a session (which
does happen in the registration matchback) we need the `VerificationCode` model
to support the matchback attributes so that we can render them in the
verification code form (hidden from the user) and submit the data back to the
app.