Support candidate sign in via the API
Updates the `SessionsController` so it behaves simiarly to the
`SignInsController` and can support the `git_api` feature being enabled.

The modelling here is a bit messy, with `Candidates::Session` and
`Candidates::VerificationCode` being similar but slightly different. Its also
complicated by the session attribute being `firstname` and not `first_name` so
we have to manually map to/from models attributes a bit.

Once the legacy code has been removed this can definately be cleaned up
further, but short of a large refactoring this is the best I can do just now to
support both CRM integration methods.