Display the correct experience type in school pages
The search results use the format_school_placement_locations helper
method to display the school experience type, so makes sense to also
use it in the school pages.

This commit also removes the format_school_experience_type helper
as it is replicating the format_school_placement_locations logic,
but without obvious use cases.