Refactor the dfe signin healthcheck
Remove the guard clause because the API Client already checks if it's
enabled before each request. Also this guard clause returns true when
the API credentials are missing which is a false positive result
for the check itself.

Return boolean value instead of a hash when the sign in works.

Rename the auth label to be more clear it's about the DfE signin api.