Remove legacy code from ContactFetcher
The `ContactFetcher` is a fairly core service and so this had a lot of knock on
effects in terms of broken specs.

In a couple of places the telephone/email being displayed has changed slightly
but I don't think its a big deal (its been this way for weeks in production --
if someone raises it as a problem we can look to modify the code that caused
the behaviour change).