Remove legacy code from Candidate and GitisAuthenticate concern
These are the two remaining core classes that tie in to a bunch of the
other code, which means numerous tests failed when pulling out the
legacy code.

I also had to pull out any remaining bits that were feature switched or
type checking in order to get everything passing.