Recreate Review Comments
ON SE we use an out of the box solution rather than a bespoke solution
we used on Get Into Teaching, when we created the one on GiT the OOB
didn't exist.  The OOB solution has an option to delete and recreate,
which will remove the old comment and add a new one at the end,
I think this should help remove the clutter of lots of comments
as well as make it clear where the comment is.