Allow non-signed in users to visit the `contact us` page
Currently, users who aren't signed in can't visit the contact us page. 

Skip authentication for this page, and conditionally set the destination and text of the button based on whether the user is signed in or not. If a user is not signed in they can go back to the main schools page, if they are signed in they can go to the dashboard.