Make sure users can complete their current sign up
I previously merged (and reverted) a change to remove date of birth from the candidate sign up. Unfortunately, we saw an exception where a user was in the middle of their sign up journey. 

Leave the date of birth attribute in the PersonalInformation model to support these users, it can be removed later.