Remove CRM from healthcheck
We have added some resilience to the candidate sign up and classroom experience note by queueing requests in the API when the CRM integration is paused.

We don't want to be alerted anymore via Slack and StatusCake when the CRM integration is paused.

Some operations will remain unavailable such as fetching a candidate (which should be a rare occurrence), but we will see these as Sentry errors.