Re-enable buttons when navigating to page with browser back button
We recently fixed some JavaScript that disables buttons so that a user can't submit a form multiple times.

Howeve, pages don't reload fully when navigating back using the browser's back button, so the button remains disabled.

We can add an event listener for the "pageshow" event to re-enable the button when this happens.