Remove non-primary phone numbers and email address
Currently, if primary phone number or primary email address has a value in the CRM, we update secondary phone number or email instead.

This was a decision made in phase 1, but following a discussion with the CRM team, we have decided that we should change the behaviour and overwrite primary phone number after a matchback, so that the contact information is always up to date (email won't actually be affected because we don't allow candidates to update their email following a matchback).

I have removed everything expect `telephone` and `email address` from the API sign up request.