Add `time_until_booking_descriptive` to reminder jobs
We want to send SMSs with either "Your school experience is tomorrow" or "your school experience is in one week" instead of the current email format which is "your school experience is in one day".

In order to use the same temple for both SMS messages, add a `time_until_booking_descriptive` string.