Use GOV.UK form builder on search form
Update search form to use GOV.UK form builder. There is no `search`
type field in the new form builder so I've updated it to be a
`text_field`; it doesn't appear to make any difference in Chrome and
I think few browsers do anything contextual for this type of field
anyway.

The labels are manually linked to the translations due to the use
of`as: ''` in the form call; the new GOV.UK form builder
doesn'tappear to handle this usage in the same way as the old one
and thetranslations don't automatically load in as you would expect.

There appears to be no error messages on this form; if you submit
with an empty location the form just reloads. The behaviour hasn't
changed here so if its something that needs addressing it can be
picked up separately.