Use GOV.UK form builder on search results page
Use the GOV.UK form builder on the search resuslts page.

The labels are manually linked to the translations due to the use of
`as: ''` in the form call; the new GOV.UK form builder doesn't
appear to handle this usage in the same way as the old one and the
translations don't automatically load in as you would expect.