Use GOV.UK form builder on confirm attendance form
Update to use the GOV.UK form builder for the confirm attendance form.

As we are using `govuk_form_with` there is no model bound to the builder and
therefore `govuk_error_summary` raises an error. As a workaround I've created a
second form with the model above that will display any errors that are present.
The errors won't link to the correct row, but they don't in master at the
moment either anyway.