Reorder presence validation to be after timeliness
At the moment if you enter an invalid date into the edit
booking date form it results in the attribute on the model
being assigned `nil` (as they date cannot be represented).

If the model is them validated we end up with a `:blank`
validation error in addition to an `:invalid_date` error. As
`govuk_error_summary` only outputs one error per attribute
we see the `:blank` one as its defined first in the model.

Instead, we can get the correct `:invalid_date` error (provided by the
Timeliness gem via a shim on ActiveRecord) if we move the presence validator to
be last. This way the timeliness error will take presedence.

In an ideal world we could use the `allow_blank: false` option as part of the
timeliness validation call, but for some reason this just doesn't appear to
behave.