Add Bookings::OutstandingTasks to easily summarise by school
At the moment we calculate the outsanding tasks for schools in
different places, using different methods. This has resulted in
the badges on the dashboard not adding up to the total that is
displayed on the school switcher page.

Extract the logic from the dashboard and encapsulte in a new
class that we can call and use in both scenarios.

The current scope for the school switcher page is more optimal
than this solution, but its impossible to try and reconcile the
logic with the dashboard as it is. This will perform individuall
requests that can then be totalled; I don't imagine the
efficiency difference will be an issue and this will be far
easier to maintain and debug.

In my testing I occasionally got badges showing on the dashboard
when there were no corresponding bookings/requests on the detail
page. I think there is overlap between
`requests_requiring_attention` and `unviewed_withdrawn_requests`
at the moment, but that will be addressed in a later PR/commit.
For now the aim is to keep it consistent with what the dashboard
is displaying.v