Add secondary subject to the CSV export
As part of the request process, candidates can select 2 subject
preferences. We only have one of these as a field int he CSV export, we
should add this as a parameter on the CSV export to be helpful for some
schools.

It looks like for the primary subject the `Booking` subject takes
presedence over the `PlacementRequest`, however we only have the
`PlacementRequest` as a reference to the secondary subject so we just
directly retreive it from there.