Fix Cucumber specs not running
The Cucumber specs are not running correctly in CI; it looks like the
`cucumber` 'mode' passes in a filter of `-t @javascript` to all the runs, which
means there are 0 matching scenarios in the regular Cucumber step and a subset
in the Chrome Cucumber step.

Also adds a `shell` mode that enables easier debugging of these kinds of issues
locally.