Only run Javascript tests with Selenium
I'm not sure why we run all the non-Javascript tests under the
`continuous_integration` profile and then all the Javascript _and_
non-Javascript tests with Selenium. I think this is redundant overlapping; we
should only be running the Javascript tests against Selenium, particularly
because they are so slow to run that way.