Simplify Selenium driver/setup
Remove `webdrivers` and instead rely on `selenium-webdriver` and `chromium` in
the Docker image.

Previously we were spinning up other instances of the application in a Selenium
hub, which is complex and I couldn't get my head around the configuration and
why it wasn't working. This makes it simpler and more inline with how we run
the RSpec tests; I don't think we're losing any value by doing it this way.