Add applied_filters method to search model
We need to be able to get a representation of the applied filters in order to
populate facet tags on the results page.

To be able to dynamically link these to the form when a user interacts with
them we will need to be able to derive the form element and value from the
field. We also need the text to display in the tag.