Update search results on tag/form change
When a user clicks to remove a tag, or clicks one of the search filters, we
want to update the search results. To make this less-jarring for the user, we
want to update the results without reloading the web page/changing their
position on the page.

Add `SearchController` to handle updating the form dynamically in response to
user interactions.

Setup Jest so that we can test our Stimulus controllers.