Search back button returns user to previous page
Instead of the back button removing applied tags we want to take the user back
to the previous page. To achieve this we can use `replaceState` instead of
`pushState`, which will update the URL but not add a new entry into the History
stack.

Also adds progressive support for search; IE11 will get the JS disabled version.