Use United Kingdom as the search region instead of England
The search region is currently set to England because School Experience is an England-only service.

However, when somebody search for a non-England area/address (e.g., Edinburgh), we receive a null result (because 'Edinburgh, England' can't be interpreted by Google Geocoding) and don't log out region and district. 

Additionally, if they are close to the border and within their chosen radius of any events, the null result means that we won't show these to the user.

Update the region to United Kingdom so that we no longer get null results.

Note: Google Geocoding is interpreting some searches okay. For example, 'LL11 5YP, England' returns the correct coordinates despite being in Wales.