Add ability to close and open a date a number of days before it starts
We are moving towards a new manage dates design. 

Part of this is to have two new fields which will set an availability window in which the day will be available to applicants - open date x number of days before & close date x number of days before.

Another part of this is to remove the active/inactive checkbox. Now, dates will become active when published, and can be made inactive by following a link in each date page.