Show targeted messages when no results from searches outside England
We want to show a message telling users that this is an England-only service, along with a link to a country specific teacher training resource.

Check the address component of the Geocoding result to see if the search was in Northern Ireland, Scotland, or Wales, and display the link and message if there are no results (there could be some results if the search is near to the border).