Pre-check all subjects for new placement dates
When a new placement date is created the user is prompted to check all subjects
that apply for the given date. A school can have many subjects and checking
them all is time consuming/an annoyance.

Instead, we want to pre-check all the subjects and let the user uncheck those
that don't apply.