Send feedback request emails on confirmation of attendance
When a school indicates that a candidate has attended a school experience
booking we want to issue an email that requests their feedback.

Add `CandidateBookingFeedbackRequest` to queue an email to the candidate.

Update `School::Attendance` to issue feedback requests for attended bookings.

Update `ConfirmAttendanceController` to load the `gitis_contact` for bookings
prior to processing the attendance; we need the contact information present so
that we can access the candidate's email.