Fix education phase seed data
When we seed the education phases they all default to `supports_subjects: true`
- this means we can never add a placement date for a primary education phase
(the control is never displayed to the user).

We seem to have manually fixed this in the production/staging database, however
when reseeding locally its an issue.

Fix by ensuring `supports_subjects` is `false` for the primary education phase.