Update Feature class to read from a JSON config
We currently have a couple of implementations of feature flags in the app (the `Feature` class and Flipper).

They aren't currently very visible, which means there are features that we don't know about (capped_bookings), or features that we though were turned on and actually weren't (sending reminder emails). 

We investigated using the Flipper dashboard for visibility, but while there is a feature (Flipper cloud) where  you can see flags across all environments, it's a paid feature.

Instead, we can use a JSON config file to manage feature flags. We will read this into a dashboard which will be accessed from any environment.

This also removes the `subject_specific_dates` flag which was enabled in all environments, and replaces the `phase` var (This was an earlier iteration of feature flagging, currently only used to disable candidate dashboard access) with feature flags instead.