Store the experience type for flex dates
to always know what experience type the candidate is going to have
or had in the past.

To make it easy for the school managers, the placement request will be
getting the experience type of the school and then the booking will be
getting it from the placement request. If however the school's
experience type is unclear(1) then the school will be asked to choose
between 'In school' and 'Virtual' when accepting a request.

(1) The school experience type is unclear when a school offers both
experiences or when the school has shifted experience types and there
are pending placement requests.