Save changes on editing individual steps when onboarded
If a school has onboarded and they edit their profile, we are finding that they
get to the preview profile page and assume their changes have been saved. They
do not realise that they need to scroll to the bottom of the profile preview
page to publish their changes.

To avoid this confusion, when a school has onboarded and they view their
profile then go on to edit one of the sections we want to save the changes as
soon as they have filled in the section (which may be multiple steps) and show
a success notification on the preview page.