Allow http connections to /metrics endpoint
When we setup Prometheus we call the app internally (bypassing the
load-balancer), however we can only do this over HTTP and not HTTPS. To allow
`/metrics` to be called over HTTP we can exclude it from the forced SSL checks.

I haven't made this change to development as we seem to force SSL via Puma so
it won't work locally.