Update school profile when not yet onboarded
Make the profile page for schools not yet onboarded match the updated design
for the onboarded school profile page.

Rename the partials to be clearer.

Remove the ability to start a placement request from the profile of a school
not yet onboarded (we no longer surface these schools in search results).

We duplicate some of the onboarded school profile template but can't reuse it
easily as it depends on the presenter, which is not set if the school has not
yet onboarded. I think this could be cleaned up in general, but leaving it for
a separate PR.