Strip PII from delayed job metric labels
The gem we use to gather delayed job metrics uses the `job.name` as a label
value; for `DelayedJob` this equates to the class name in addition to the
payload. Some of our jobs (such as emails/sms) have sensitive data in the
payload, and so it can end up in our metrics instance.

To avoid this, we can monkey patch the `labelize` method of the gem and
overwrite the `worker` value to use the `Delayed::Job` class name, which
doesn't include PII but will still let us diferentiate between jobs.

I tried to test this via the plugin behaviour but for some reason I wasn't able
to mock out the `::Yabeda.delayed_job.jobs_enqueued_total` attribute; it was
never called. Instead Ive opted to test the patch directly.