Exclude IE 11 in the browserlist
Stimulus controllers are failing with the following error:
Uncaught (in promise) TypeError: class constructors must be invoked with 'new'"


It seems that Stimulus 3 has dropped IE 11 support and uses ES6 as the 
compile target, which cause the issue. Excluding IE 11 from the browserlist
fixes the issue.

A very small percentage of our users are using IE 11, so we are happy with them
to have a degraded but still a very functional experience.