Spike on-boarding with task progress
We want to perform user testing on a new on-boarding design which - instead of
being a linear flow of forms - has an overview/progress page that the user
keeps returning to after completing each section. This way they get a clear
view of how far through the process they are.