Migrate to Shakapacker (#2345)
* Migrate to webpack 5

This commit
- Bumps up webpack to v5.69.1
- Replaces webpacker with shakapacker v6.1.1
- Updates the webpacker.yml config
- Moves the browserslist in the package.json and deletes the redundant
  dot file
- Uses a single webpack config. Doesn't seem we need a config per env
- Moves the extensions from webpacker to webpack config. Also removes
  unused extensions
- Adds webpack css and scss loaders
- Renames the webpack binstubs to webpacker
- Updates the paths of the static files in the application layout

* Call the javascript_pack_tag in the application layout at most once

Shakapacker complains when javascript_pack_tag is called multiple times.

Refactor the gtm pack to instantiate gtm only when it is enabled.

* Exclude IE 11 in the browserlist

Stimulus controllers are failing with the following error:
Uncaught (in promise) TypeError: class constructors must be invoked with 'new'"

It seems that Stimulus 3 has dropped IE 11 support and uses ES6 as the 
compile target, which cause the issue. Excluding IE 11 from the browserlist
fixes the issue.

A very small percentage of our users are using IE 11, so we are happy with them
to have a degraded but still a very functional experience.