Add ability to review recurrences
After we've selected the recurring criteria we show the next 6 months of dates
(from the day after the placement date) so that the user can select/deselect
additional dates before continuing.

I've not mapped the dates back to a model for validation as we only list out
valid dates for them to select. They _could_ hack the form to submit an invalid
date, but at the end of the day its just not going to show up so wouldn't cause
an issue.