Add final step to publish dates
Add a new, final step to the form where you can review and change details and
then publish the date and all recurrences.

Due to the way the configuration step works we need to be able to differentiate
in order to know when a user has passed this step. I've added a 'publishable`
attribute to `PlacementDate` for this purpose; signifying the user has
sufficiently filled out the placement date form and is now able to go back and
edit their changes.