Prevent placement dates from being created on a weekend
We want to prevent placement dates from being created on a weekend;
this is more of an issue now we generate recurring dates. It should
hopefully prevent user-error and simplifies the following steps
around recurring dates.

Initially limits the validation to unpublished placement dates in
case we have some already in the system (we don't want them to
become invalid all of a sudden).