Reduce DelayedJob sleep_delay
The `sleep_delay` is currently 15 seconds, which means it can take
up to 30 seconds for a queued job to be picked up and processed;
this is fairly noticable with the registration confirmation emails.

Reducing this time will make the jobs execute more promptly and
isn't going to have an impact on resource consumption (each time a
worker checks it'll query the database, but GSE sits fairly idle
most of the time anyway so we have plenty of capacity).