Split school profile into tabs
As part of a design refresh, we have decided to revisit the profile updating page (before we do anything more radical).

The current page is rather long, so we're splitting it according to the <h2/3> tags. This design takes elements of the Publish service, so there's some evidence to suggest this will work.