Minor copy tweaks based on feedback
Update the `placement_details` step to pluralise some of the
headings when the user has opted-in to recurring dates.

Remove the date heading from the `configurations` step and switch
the questions around so the "experience type" question is first and
becomes the main heading.

Minor tweaks to the summary rows on the review step to account for
recurring dates. We now also list out all individual recurrences on
the review step.