Add total counts to school search filters
We want to be able to indicate to the user if there are going to be
any results for a given search filter.

Add counts to each filter; if they are in the same group they are
OR'ed together, whereas different groups of filters are AND'ed
together.

Remove the brackets from the phase labels to avoid doubling up, e.g.
"Primary (11 - 14) (2)".