Change CF healthcheck to use port
We changed the health check to `process` when we were struggling to
get the DelayedJobs metrics working a few weeks ago. This has resulted
in a few seconds of downtime during deployments to production as the
process goes 'live' before the web service has fully booted;
StatusCake then hits the app and it gets an error. I don't think the
`process` check type was necessary in the end for the metrics to work
so reverting back to `port`.