Fix view specs not resolving templates
I'm not sure why but they don't resolve if we specify the `.html.erb`
extension; omit it and it and RSpec/ActionView seems to find it fine.

The robots template was also failing to resolve in the specs and when running
the app; again I'm not sure whats changed but using `text` instead of `txt` in
everything apart from the URL seems to fix it. It may be something around the
MIME types that has changed.