Placate Bullet query checker
Bullet is complaining about needing to eager-load a few relationships on
various queries. I'm not sure why its started complaining about these with the
Rails 7 updated; I suspect one is to do with chaning `includes` statements and
the other may be around including the subject of the query so I re-wrote it
(instead of `current_candidate.placement_requests` and trying to eager-load
`candidate` we query the base model we want and supply the `candidate` as a
conditional on the `where`).