Move basic auth config into controllers
The autoloader appears to run after the initializers in Rails 7 so we were
getting uninitialized constant errors for the `ApplicationController` and
`HealthcheckController`. To avoid this we can move the logic into the
controllers themselves instead of having it in a separate initializer. Also
adds test coverage for the basic auth functionality.