Ensure step is invalid initially unless pre-filled
When a `SchoolProfile` is not yet persisted the fee attributes are `nil`; we
want to ensure the step is always shown (invalid) when this is the case.

When a school edits a profile these fields will be explicitly `false` so we can
pre-select `None` if they are all `false` (meaning the step will be valid and
skipped unless edited directly).