Correct flow when school chooses DBS check required
When a school chooses to update their DBS check requirements to yes, we want to redirect them to the fees step, as they won't yet have had the option to select whether any DBS fees are required.

An an attribute to the fees model in order to control the flow in this scenario.