Add a "none" option to the candidate dress code requirements
GOV guidance says that we should have an option for "none", when that is a valid answer.

The built in GOV "none" checkbox requires the fields to have matching names, so introduce a new "selected_requirements" attribute, which we can map the user's selection to.

We also need to make sure "none" works with JavaScript disabled, so add some addition validation.