Ensure fees step is displayed if DBS check required
We need to make sure that if a school sets up their profile then edits it and
changes DBS check to be required that they are then taken to the fees step
again to optionally select if DBS fees apply.

This involves updating the `CurrentStep#fees_required?` logic to return `true`
if a DBS check is required and the DBS fees have not yet been specified (as
they are omitted from the fees step when a DBS check is not required).

We can tell if DBS fees have been specified by querying the underlying
`dbs_fees` attribute of the `SchoolProfile` which will be explicitly `nil` if
the user has not yet been presented with this option.

We need to default the `dbs_fees_specified` attribute to `true` to ensure that
the user can progress past the fees step on their initial run through the
on-boarding process.