Add concern to encrypt ActiveModel attributes
We plan on encrypting the attributes of a form model when rendering the GET
request that contains PII in the URL query string. In order to do this we need
to be able to query the encrypted attributes of a model (to construct the URL)
and then be able to instantiate the same model using the encrypted attributes
(decrypting the fields into the model attributes).

The usage will be along the lines of:

```
model = Model.new(attributes_from_post_request)
url = model_path(model.encrypted_attributes)
redirect_to url
model = Model.new_decrypt(attributes_from_get_request)
```