Encrypt postcode in school search
Encrypt the location in the school search URL so as not to expose it to Google
Analytics/the browser history.

In order to do this we include the `EncryptedAttributes` concern into the
`Events::Search` model and change the form method from `get` to `post`. This
ensures the location is transmitted securely to the controller action, whereby
we can encrypt it and redirect to the `GET` search endpoint, decrypt it and
display the relevant results to the user. This method ensures the location is
not leaked to the URL and yet the search results URL can be copied/bookmarked.

Update search model to use `ActiveModel::Attributes`.