Add none option to candidate requirements step
We want to have a 'None' option on the candidate requirements checkbox for if
none of the other options apply. The form builder requires the checkboxes to
have the same attribute name in order to add a `None` option.

Refactor `CandidateRequirementsSelection` to have a single
`selected_requirements` attribute that contains the selected options.