Add rake task to transfer a placement request to another school
When a school becomes an academy they end up with a new School Experience
account and often get locked out of their old account/no longer sign in to it.
They may have pending placement requests that they have not yet actioned and
want to see them under the new school account that they use.

Add a rake task to transfer a placement request from one school to another
(unlinking it from the original school and linking it to the new school).
Currently the functionality is limited/restricted and it will only transfer new
placement requests (providing matching subjects and placement dates can be
found).

If we find we need to transfer placement requests in other states we can expand
the transferring library to accommodate.