Add rake task to identify on-boarded, closed schools and their replacements
A school that has been on-boarded to School Experience may subsequently be
closed (according to the latest GIAS data). It could also be replaced by a new
school (as is the case when a school transitions to an academy). We need a way
to identify on-boarded schools that are now closed and, if possible, find the
URN of their replacement school.

Add library to reconcile our data set with GIAS and a corresponding rake task
to output on-boarded, closed schools and their replacement school URN(s).