Periodically email list of onboarded, closed schools
We want to keep track of schools that have been onboarded to the service but
subsequently marked as closed according to the latest GIAS data. We will aim to
get in touch with these schools and transition them to a new account on the
system (for their replacement school if one exists).

Run periodic cron job (Mondays at 9am) to email the service inbox with a list
of closed, on-boarded school details.

Originally I wanted to include this as a markdown formatted table but the
GOV.UK Notify service is very restricted on what it will accept and it blocks
this kind of content. Instead I've opted to format it as a sentence that can be
read out (I wasn't even able to get line breaks working consistently).

With hindsight this may have been better as a Slack integration, but to do that
we would either need to expose our production credentials to our CI pipeline or
pull in a whole new Slack integration, which felt like overkill.