Ensure candidate requirements options share common legend
Its not clear to users with a screen reader that these are grouped; they should
have a common legend.

We also have this issue on the `key_stage_lists` and `phases_lists` forms,
however as they are checkboxes for individual named attributes we can't easily
group them, so I've left that until we refactor those steps to be inline with
`candidate_requirements` (they should also have a `None` option).