Return focus to live filter after search
When we perform a 'live search' the results, including filters, are overwritten
with the response from the server. This results in the loss of focus state on
the page and tabbing will continue back at the top of the document.

Instead, we want the focus to return to the filter element that caused the live
update.

Pulls in `jest-environment-jsdom` - I think this used to be bundled with
`jest`.