Include number of results in page title
The search results page was flagged as having a duplicate page title after
executing a search as the results change but its potentially not clear to the
user that they are now on a new/updated page.

Update page title to include the number of results. Ensure the page title
updates on live search updates.