Prevent InvalidAuthenticityToken errors in Safari
There is an issue with CSRF tokens and mobile Safari whereby the page can be
reloaded from the store/cache and submitted after the session has been cleared
(resulting in an `InvalidAuthenticityToken` error). It's not enough to specify
`max-age=0`; we need to tell Safari to not store the page at all with
`no-store, no-cache` to ensure the page is always retrieved from the server and
has a valid CSRF token.